@@ -12,32 +12,37 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.9",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/platform-server": "^5.0.0",
-    "@angular/router": "^5.0.0",
-    "@ng-bootstrap/ng-bootstrap": "^1.0.0-beta.5",
+    "@angular/animations": "^6.1.7",
+    "@angular/cdk": "^6.2.0",
+    "@angular/common": "^6.1.7",
+    "@angular/compiler": "^6.1.7",
+    "@angular/core": "^6.1.7",
+    "@angular/forms": "^6.1.7",
+    "@angular/http": "^6.1.7",
+    "@angular/material": "^6.4.7",
+    "@angular/platform-browser": "^6.1.7",
+    "@angular/platform-browser-dynamic": "^6.1.7",
+    "@angular/platform-server": "^6.1.7",
+    "@angular/router": "^6.1.7",
+    "@ng-bootstrap/ng-bootstrap": "^3.2.0",
     "core-js": "^2.4.1",
     "font-awesome": "^4.7.0",
+    "global": "^4.3.2",
     "ng-contenteditable": "^1.0.5",
     "primeng": "^5.2.0-rc.2",
-    "rxjs": "^5.5.2",
-    "zone.js": "^0.8.14"
+    "rxjs": "^6.3.2",
+    "rxjs-tslint": "^0.1.5",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.5.0",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/language-service": "^4.2.4",
+    "@angular-devkit/build-angular": "~0.8.0",
+    "@angular/cli": "^6.2.2",
+    "@angular/compiler-cli": "^6.1.7",
+    "@angular/language-service": "^6.1.7",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
-    "codelyzer": "~3.1.1",
+    "codelyzer": "^4.4.4",
     "jasmine-core": "~2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
     "karma": "~1.7.0",
@@ -50,6 +55,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~3.2.0",
     "tslint": "~5.3.2",
-    "typescript": "2.4.2"
+    "typescript": "^2.7.2"
   }
 }
