@@ -2,8 +2,7 @@ import {Injectable} from '@angular/core';
 import {User} from '../api/user';
 import {Item} from '../api/item';
 import {ITEMS} from '../mock.data';
-import {Observable} from 'rxjs/Observable';
-import {Subject} from 'rxjs/Subject';
+import {Observable, Subject} from 'rxjs';
 
 @Injectable()
 /**
